fixing bux
There was a bug where the app would not load on phones. It was due to CSS - some attributes do not properly work on phones while the app runs on Web version. There is still a bug related to not being able to click on an event to view event's description. Working on it right now.